+++ packages/adapter-fcm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @msgly/fcm
 
+## 1.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @msgly/core@1.6.0
+
 ## 1.5.0
 
 ### Patch Changes